Covariance Matrix: Added unit test based on a signal gernerated by the Oscillator class
Creates an array of random phase shift values and assigns it to the oscillator channels. The covariance matrix gets calculated based on the signals generated by the oscillator. According to the theory it should contain the phase differences between channel 0 and the other channels in its first row, so the Unit test checks for equality of those values